Finished CRUD Functionality Redux
Was able to get all CRUD working and finished everything in redux. Bugs may occur if trying to enter smurf forms before CDM from the root directory routed  component finishes running. Solution to this would be to replace the NavLink for it with a spinner Loader until the CDM is completed. Had a few blockers, i was not properly forwarding the updateSmurf method. My id checks were not working because the original data had a smurf with no ID property on the object.